backend/pyproject.toml
[project]
name = "ai-fitness-copilot"
version = "0.1.0"
description = "Hybrid fitness analysis backend with deterministic nutrition logic and insight generation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
    "openai>=1.3.0"
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
]

[build-system]
requires = ["setuptools>=80.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]